requirements.txt
Flask==3.1.2
flask_mail==0.10.0
fpdf2==2.8.5
itsdangerous==2.2.0
mysql_connector_repackaged==0.3.1
Werkzeug==3.1.5
